# [clang] Optimize castToDeclContext for 2% improvement in build times