@Destroyerrrocket 2024-01-04
Addressed the first review comments! Thank you for your time, I appreciate it :)